+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "IPv6 shortener".
 - Tests for "Zhiwei Sun squares".
 - Tests for "Stall tilt".
+- Tests for "What's so complex about Mandelbrot?".
 
 ## [1.9.0] - 2022-06-01
 ### Added